feat: implementing Vault architecture

+++ contracts/oracle/PriceConsumerV3.sol
@@ -4,8 +4,8 @@ pragma solidity ^0.6.6;
 import "@chainlink/contracts/src/v0.6/interfaces/AggregatorV3Interface.sol";
 
 contract PriceConsumerV3 {
-
     AggregatorV3Interface internal priceFeed;
+
     /**
      * Network: Kovan
      * Aggregator: ETH/USD
@@ -18,14 +18,17 @@ contract PriceConsumerV3 {
     /**
      * Returns the latest price
      */
-    function getLatestPrice() public view returns (int) {
+    function getLatestPrice() public view returns (int256) {
         (
-            /* uint80 roundID */,
-            int price,
-            /* uint startedAt */,
-            /* uint timeStamp */,
-            /* uint80 answeredInRound */
-        ) = priceFeed.latestRoundData();
+            ,
+            /* uint80 roundID */
+            int256 price, /* uint startedAt */
+            ,
+            ,
+
+        ) = /* uint timeStamp */
+        /* uint80 answeredInRound */
+        priceFeed.latestRoundData();
         return price;
     }
 }
